- Added function to renderer that returns a const ref to current render batch - Added .cso files to the .gitignore - Added pragma warning ignores for loop unrolling in shaders - Added EntityStatic class to replace the missing TestEntity - Fixed ALL warnings by casting/matching comparison types - Removed classdiagrams I added by accident initially - Removed old .orig files from merges past - Now perfoms memset(0) whenever a new Vertex struct is initialized to prevent invalid values (now all go to 0 instead of a "random" number) - Added CalculateTangents to FBX import as a temporary measure until tangents importing is implemented

+++ StateManager.cpp
@@ -36,7 +36,7 @@ void StateManager::SetState(GameState newState)
 
 void StateManager::LoadScene(Scene * scene)
 {
-	for (int i = 0; i < scene->GetEntityList().size(); i++)
+	for (size_t i = 0; i < scene->GetEntityList().size(); i++)
 	{
 		Renderer::Instance()->StageEntity(scene->GetEntityList()[i]);
 	}
@@ -45,7 +45,7 @@ void StateManager::LoadScene(Scene * scene)
 
 void StateManager::UnloadScene(Scene * scene)
 {
-	for (int i = 0; i < scene->GetEntityList().size(); i++)
+	for (size_t i = 0; i < scene->GetEntityList().size(); i++)
 	{
 		Renderer::Instance()->UnstageEntity(scene->GetEntityList()[i]);
 	}
